Change Type for analysis data

+++ analysis/main.cc
@@ -7,7 +7,7 @@ Napi::Promise SumAsyncPromise(const Napi::CallbackInfo& info) {
 	uint32_t samples = info[0].As<Napi::Number>().Uint32Value();
 
 	Napi::Uint16Array deltas = info[2].As<Napi::Uint16Array>();
-	Napi::Float32Array values = info[1].As<Napi::Float32Array>();
+	Napi::Uint16Array values = info[1].As<Napi::Uint16Array>();
 
 	std::cout << "processing data..." << std::endl;
 	std::cout << "samples: " << samples   << std::endl;
